Have DDLogMessageOptions use NSInteger instead of Int.

+++ Classes/DDLog.h
@@ -535,7 +535,7 @@ NSString * DDExtractFileNameWithoutExtension(const char *filePath, BOOL copy);
  * If you write custom loggers or formatters, you will be dealing with objects of this class.
  **/
 
-typedef NS_OPTIONS(int, DDLogMessageOptions) {
+typedef NS_OPTIONS(NSInteger, DDLogMessageOptions) {
     DDLogMessageCopyFile     = 1 << 0,
     DDLogMessageCopyFunction = 1 << 1
 };
